+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Folding@home Web Control Changelog
  - Removed random sort of news articles.
  - Get news articles from new API endpoint.
  - Added log download log feature.
+ - Color background of log lines in light mode, re:#291
 
 ## v8.5.3
  - Added user timeseries charts